requirements.txt
kivy[base]
kivy_examples
GitPython